% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decluster.R
\name{decluster}
\alias{decluster}
\title{A declustering scheme for dependent exceedances}
\usage{
decluster(x, index, uq = 0.9, r = 0, u, doplot = TRUE, ...)
}
\arguments{
\item{x}{Numeric, sequence of depedent random variables.}

\item{index}{Numeric, vector specifying the index for the elements of x.}

\item{uq}{Numeric in (0, 1), calculates the threshold as quantile(x, uq).}

\item{r}{Numeric, nonnegative integer, the decluster parameter.}

\item{u}{Numeric, if specified, the threshold is u and overrides what is
chosen for uq.}

\item{doplot}{Logical, should a plot be generated?}

\item{...}{Additional arguments passed to the plot that is generated if
doplot = TRUE.}
}
\value{
The maximum _excess_ for each cluster, the number of clusters, the number of exceedances,
the number of observations, the threshold used, the decluster parameter
}
\description{
Calculate clusters of exceedances for a sequence of dependent random variables
}
\details{
The argument index may be used to specify when each observation in x has occurred.
For example, suppose x contains daily observations in January on two separate years,
making length(x) = 62. Leaving index blank would treat January 31 of year one as
the day just before January 1 of year two, allowing the possibility of clusters to
exist that span both year. In this example, index should be c(1:31, 366:396).

The decluster parameter r controls the cluster sizes. Two exceedances are part of
different clusters if there are at least r non-exceedances between the two obsverations.
This has the consequence that for r = 0, every observation will be its own cluster.

When multiple observations are part of the same cluster, the maximum value in that
cluster is the value that is return.

If doplot = TRUE (the default), then a plot of (index, x) is shown with a horizontal line
at the threshold, and clusters are marked by the shape and color. (Note: the color/shape
pattern of clusters is repeated every 77 clusters).
}
\examples{
set.seed(1)
x = rnorm(100)
decluster(x, uq = 0.5, r=0) # 50 clusters
decluster(x, uq = 0.5, r=1) # 26 clusters
decluster(x, uq = 0.5, r=2) # 12 clusters
decluster(x, uq = 0.5, r=3) # 4 clusters
decluster(x, uq = 0.5, r=4) # 4 clusters
decluster(x, uq = 0.5, r=5) # 3 clusters
decluster(x, uq = 0.5, r=6) # 2 clusters
decluster(x, uq = 0.5, r=7) # 1 cluster
}
