% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mrl.R
\name{plot_mrl}
\alias{plot_mrl}
\title{Generate a mean residual life plot}
\usage{
plot_mrl(x, qs = c(0, 0.5, 0.75, seq(0.8, 1, by = 0.01)), noopt = FALSE,
  ...)
}
\arguments{
\item{x}{Numeric, vector.}

\item{qs}{Numeric, vector quantiles which will be marked on the x-axis. This
does not affect the computation of the mean residual life.}

\item{noopt}{Logical, should the plot function be called with no additional options?
Defaults to FALSE. This would need to be TRUE for full control over the plot.}

\item{...}{Additional arguments passed to plot().}
}
\description{
Plots the mean residual life for a vector of data, with 95\% confidence bands
}
\details{
If the generalized Pareto is a valid model for the tail of a distribution (i.e. given
a threshold), then the mean residual life plot should be approximately linear beyond
the threshold.

The motivation for using this plot is in threshold selection. By using the plot, one
can observe the lowest threshold for which the generalized Pareto may be valid.

If noopt = TRUE, anything regarding the plot can be specified by the user except ylim
and the lines.
}
\examples{
plot_mrl(rnorm(1000))
plot_mrl(rgamma(1000, 1, 1))
}

