% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_hier.R
\name{theta_hier}
\alias{theta_hier}
\title{Extremal index estimation}
\usage{
theta_hier(y, u, n, prior, likelihood, K = 1, ...)
}
\arguments{
\item{y}{Numeric, sequence of depedent random variables. May be a matrix.
See details.}

\item{u}{Numeric, the threshold. Defaults to quantile(y, 0.90).}

\item{prior}{List.}

\item{likelihood}{Character, either "ferro" or "suveges" (default).}

\item{K}{Numeric, the run parameter for likelihood = "suveges" only. Defaults
to K = 1, the model proposed in Suveges (2007).}

\item{...}{Additional arguments passed to mwBASE::mcmc_sampler.}

\item{ord}{Numeric, vector of length R = NCOL(y). Defaults to 1:R. See details.}
}
\description{
Fill
}
\details{
The data y is expected to be observed values along an evenly-spaced grid of width 1.
If y is an n x R matrix, then it will be collapsed to y = c(y[,ord]). This may be
the case when working with multiple realizations from the same process or with
specific seasons on a time-series.

The 'classical' option for method produces the estimates given by either Ferro and
Segers (2003) or Suveges and Davison (2010). When method == 'bayesian' the likelihoods given in
the aforementioned papers are used to produce posterior samples for theta.
}
