% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta_uni.R
\name{theta_uni}
\alias{theta_uni}
\title{Extremal index estimation based on Ferro and Segers (2003)}
\usage{
theta_uni(y, u, ord, prior, ...)
}
\arguments{
\item{y}{Numeric, sequence of depedent random variables. May be a matrix.
See details.}

\item{u}{Numeric, the threshold. Defaults to quantile(y, 0.90).}

\item{ord}{Numeric, vector of length R = NCOL(y). Defaults to 1:R. See details.}

\item{prior}{List.}

\item{...}{Additional arguments passed to mwBASE::mcmc_sampler.}
}
\description{
Uses the likelihood provided in Ferro and Segers (2003) in a Bayesian setting
to obtain posterior samples of the extremal index. Separates exceedances into
clusters.
}
\details{
The data y is expected to be observed values along an evenly-spaced grid of width 1.
If y is an n x R matrix, then it will be collapsed to y = c(y[,ord]). This may be
the case when working with multiple realizations from the same process or with
specific seasons on a time-series.
}
\examples{
### Generate dependent random variables with standard Frechet marginals
set.seed(1)
n = 365*50     # 50 years of data, no leap years

theta = 0.25
W = -1/log(runif(n))
y = double(n)
y[1] = W[1] / theta
for (i in 2:n)
    y[i] = max((1-theta)*y[i-1], W[i])

out = theta_uni(y, quantile(y, 0.95), nburn = 40000, nmcmc = 20000)

mean(out$mcmc)     # 0.277
quantile(out$mcmc, c(0.025, 0.975))    # c(0.248, 0.306)


### Generate dependent random variables with standard Frechet marginals,
### but only for winter, say.
set.seed(1)
R = 50   # number of seasons (years)
n = 90   # number of observations (days) per season

theta = 0.25
W = matrix(-1/log(runif(n*R)), n, R)
y = matrix(0, n, R)
y[1,] = W[1,] / theta
for (i in 2:n)
    y[i,] = apply(rbind((1-theta)*y[i-1,], W[i,]), 2, max)

out = theta_uni(y, quantile(y, 0.95), nburn = 40000, nmcmc = 20000)

mean(out$mcmc)     # 0.267
quantile(out$mcmc, c(0.025, 0.975))    # c(0.214, 0.325)
}

