% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_excess.R
\name{uni_excess}
\alias{uni_excess}
\title{Univariate threshold excess model for climate observations and simulations}
\usage{
uni_excess(x, uq = 0.9, threshold, r = 0, m.ksi = 0, s.ksi = 10,
  nburn = 10000, nmcmc = 10000, window = 200, chain_init)
}
\arguments{
\item{x}{character, the name of the data, e.g. "decadal1961",
"historical", or "obs19501999", corresponding to a
processed data set.}

\item{uq}{numeric, within (0, 1), the quantile of the complete data
(includes replicates if available) that determines the
threshold}

\item{threshold}{numeric, values greater than threshold are considered
excess. Defaults to the threshold given by uq, but if
specified, uq is ignored.}

\item{r}{numeric, nonnegative integer, specifies the degree
of separation between clusters in a declustering
scheme. A value of r=0 means each excess is given
its own cluster. If r is not specified, it will be
selected automatically.}

\item{m.ksi, s.ksi}{numeric, prior mean and sd for ksi, ksi is normal}

\item{nburn}{numeric, nonnegative integer, the first nburn samples
are discared. Defaults to 80000.}

\item{nmcmc}{numeric, positive integer, the number of samples that are
kept _post_ burn-in. That is, a total of nburn + nmcmc
iterations of the chain are run. Defaults to 40000.}

\item{window}{numeric, during the burn-in, the covariance matrix of the
proposal distribution is changed every window iterations
to improve the acceptance rate. Also, the iteration number
is displayed every window iterations. Defaults to 500.}

\item{chain_init}{numeric vector of length 2, specifying the starting
location of the Markov chain for (sigma, ksi). Defaults to
c(1, 1e-6).}
}
\description{
Fit a univariate threshold exceedance model for a processed data set. The
data should be in a matrix format where the first column is the time and
all subsequent columns contain the values of the time series.
}
\details{

}

