% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uni_excess.R
\docType{data}
\name{x}
\alias{x}
\title{Univariate threshold excess model}
\format{An object of class \code{numeric} of length 10000.}
\usage{
x
}
\arguments{
\item{x}{Numeric vector of data.}

\item{uq}{numeric, within (0, 1), the quantile of the complete data
(includes replicates if available) that determines the
threshold}

\item{threshold}{numeric, values greater than threshold are considered
excess. Defaults to the threshold given by uq, but if
specified, uq is ignored.}

\item{r}{Numeric, nonnegative integer, specifies the degree
of separation between clusters in a declustering
scheme. Defaults to r = 0 (no clustering).}

\item{m.ksi, s.ksi}{numeric, prior mean and sd for ksi, ksi is normal}

\item{nburn}{numeric, nonnegative integer, the first nburn samples
are discared. Defaults to 80000.}

\item{nmcmc}{numeric, positive integer, the number of samples that are
kept _post_ burn-in. That is, a total of nburn + nmcmc
iterations of the chain are run. Defaults to 40000.}

\item{window}{numeric, during the burn-in, the covariance matrix of the
proposal distribution is changed every window iterations
to improve the acceptance rate. Also, the iteration number
is displayed every window iterations. Defaults to 500.}

\item{chain_init}{numeric vector of length 2, specifying the starting
location of the Markov chain for (sigma, ksi). Defaults to
c(1, 1e-6).}
}
\description{
Fit a standard univariate threshold exceedance model.
}
\details{

}
\keyword{datasets}
